Lab Variables
 Labs are not turned in via Canvas. Instead the are checked off in class. Alternatively you can email me a recording of your lab(s).

They are posted on Canvas with a due date, because this allows you to see them listed in order when you look up your grades. However, the posted date isn't really a due date. It just indicates, when we introduced a lab in class.

You have some flexibility when it comes to checking off your labs. However, make sure to check them off before midterms and final prep. That's when they 'expire'.

That being said, you are strongly encouraged to complete labs early because they help you practice concepts that you need for your assignments.